market_screener/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click-spinner==0.1.10
colorama==0.4.4
cssselect==1.1.0
dnspython==2.2.1
exchange-calendars==3.6.3
fake-useragent==0.1.11
feedparser==6.0.8
fire==0.4.0
idna==3.3
importlib-metadata==4.11.3
korean-lunar-calendar==0.2.1
lxml==4.8.0
multitasking==0.0.10
numpy==1.22.3
pandas==1.4.2
pandas-market-calendars==3.4
pandas-ta==0.3.14b0
parse==1.19.0
psycopg2==2.9.3
pyee==8.2.2
pyluach==2.0.0
pymongo==4.1.1
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
termcolor==1.1.0
toolz==0.11.2
tqdm==4.64.0
urllib3==1.26.9
w3lib==1.22.0
websockets==10.3
yahoo-fin==0.8.9.1
yfinance==0.1.70
zipp==3.8.0
